+++ examples/dagster_examples/toys/stdout_spew.py
@@ -37,5 +37,5 @@ def spew(_, name):
 @pipeline(mode_defs=[ModeDefinition(executor_defs=default_executors)])
 def stdout_spew_pipeline():
     out_1, out_2 = spawn()
-    spew.alias('spew_1')(name=out_1)
-    spew.alias('spew_2')(name=out_2)
+    spew(name=out_1)
+    spew(name=out_2)