+++ python_modules/dagster/dagster/core/execution/compute_logs.py
@@ -1,6 +1,7 @@
+from __future__ import print_function
+
 import io
 import os
-import signal
 import subprocess
 import sys
 import time
@@ -10,6 +11,8 @@ from dagster import check
 from dagster.core.execution.context.system import SystemStepExecutionContext
 from dagster.core.storage.compute_log_manager import ComputeIOType
 from dagster.utils import ensure_file
+
+WIN_PY36_COMPUTE_LOG_DISABLED_MSG = '''\u001b[33mWARNING: Compute log capture is disabled for the current environment. Set the environment variable `PYTHONLEGACYWINDOWSSTDIO` to enable.\n\u001b[0m'''
 
 
 @contextmanager
@@ -28,19 +31,41 @@ def mirror_step_io(step_context):
     )
 
     manager.on_compute_start(step_context)
-    with mirror_stream(sys.stderr, errpath):
-        with mirror_stream(sys.stdout, outpath):
-            # compute function executed here
-            yield
+    with mirror_io(outpath, errpath):
+        # compute function executed here
+        yield
     manager.on_compute_finish(step_context)
 
 
+def should_disable_io_stream_redirect():
+    # See https://stackoverflow.com/a/52377087
+    return (
+        sys.platform == 'win32'
+        and sys.version_info.major == 3
+        and sys.version_info.minor >= 6
+        and not os.environ.get('PYTHONLEGACYWINDOWSSTDIO')
+    )
+
+
+def warn_if_compute_logs_disabled():
+    if should_disable_io_stream_redirect():
+        print(WIN_PY36_COMPUTE_LOG_DISABLED_MSG)
+
+
 @contextmanager
-def mirror_stream(stream, path, buffering=1):
+def mirror_io(outpath, errpath, buffering=1):
+    with mirror_stream(outpath, ComputeIOType.STDOUT, buffering):
+        with mirror_stream(errpath, ComputeIOType.STDERR, buffering):
+            yield
+
+
+@contextmanager
+def mirror_stream(path, io_type, buffering=1):
     ensure_file(path)
-    with tailf(path):
+    from_stream = sys.stderr if io_type == ComputeIOType.STDERR else sys.stdout
+    with tailf(path, io_type):
         with open(path, 'a+', buffering=buffering) as to_stream:
-            with redirect_stream(to_stream=to_stream, from_stream=stream):
+            with redirect_stream(to_stream=to_stream, from_stream=from_stream):
                 yield
 
 
@@ -51,7 +76,7 @@ def redirect_stream(to_stream=os.devnull, from_stream=sys.stdout):
     from_fd = _fileno(from_stream)
     to_fd = _fileno(to_stream)
 
-    if not from_fd or not to_fd:
+    if not from_fd or not to_fd or should_disable_io_stream_redirect():
         yield
         return
 
@@ -70,42 +95,77 @@ def redirect_stream(to_stream=os.devnull, from_stream=sys.stdout):
             os.dup2(copied.fileno(), from_fd)
 
 
+POLLING_INTERVAL = 0.1
+
+
 @contextmanager
-def tailf(path):
-    if sys.platform == 'win32':
-        # no tail, bail
+def tailf(path, io_type=ComputeIOType.STDOUT):
+    # Cannot use multiprocessing here because we already may be in a daemonized process
+    # Instead, invoke a thin wrapper around tail_polling/tail_posix using the dagster cli
+    cmd = 'dagster utils tail {} --parent-pid {} --io-type {}'.format(
+        path, os.getpid(), io_type
+    ).split(' ')
+    tail_process = subprocess.Popen(cmd)
+
+    try:
         yield
-    else:
-        # unix only
-        cmd = 'tail -F -n 1 {}'.format(path).split(' ')
+    finally:
+        if should_compute_log_tail_poll():
+            time.sleep(2 * POLLING_INTERVAL)  # allow tail to belch output before killing it
+        tail_process.terminate()
 
-        # open a subprocess to tail the file and print to stdout
-        tail_process = subprocess.Popen(cmd)
 
-        # fork a child watcher process to sleep/wait for the parent process (which yields to the
-        # compute function) to either A) complete and clean-up OR B) segfault / die silently.  In
-        # the case of B, the spawned tail process will not automatically get terminated, so we need
-        # to make sure that we terminate it explicitly and then exit.
-        watcher_pid = os.fork()
+def should_compute_log_tail_poll():
+    return sys.platform == 'win32'
 
-        if watcher_pid == 0:
-            # this is the child watcher process, sleep until orphaned, then kill the tail process
-            # and exit
-            while True:
-                if os.getppid() == 1:  # orphaned process
-                    time.sleep(1)
-                    tail_process.terminate()
+
+def tail_polling(filepath, stream=sys.stdout, parent_pid=None):
+    '''
+    Tails a file and outputs the content to the specified stream via polling.
+    The pid of the parent process (if provided) is checked to see if the tail process should be
+    terminated, in case the parent is hard-killed / segfaults
+    '''
+    with open(filepath, 'r') as file:
+        for block in iter(lambda: file.read(1024), None):
+            if block:
+                print(block, end='', file=stream)
+            else:
+                if parent_pid and current_process_is_orphaned(parent_pid):
                     sys.exit()
-                else:
-                    time.sleep(1)
-        else:
-            # this is the parent process, yield to the compute function and then terminate both the
-            # tail and watcher processes.
-            try:
-                yield
-            finally:
+                time.sleep(POLLING_INTERVAL)
+
+
+def tail_posix(filepath, stream=sys.stdout, parent_pid=None):
+    '''
+    Tails a file and outputs the content to the specified stream via unix command `tail` (inotify).
+    The pid of the parent process (if provided) is checked to see if the tail process should be
+    terminated, in case the parent is hard-killed / segfaults
+    '''
+    cmd = 'tail -F {}'.format(filepath).split(' ')
+    tail_process = subprocess.Popen(cmd, stdout=stream)
+    try:
+        while True:
+            if parent_pid and current_process_is_orphaned(parent_pid):
                 tail_process.terminate()
-                os.kill(watcher_pid, signal.SIGTERM)
+                sys.exit()
+            time.sleep(1)
+    except KeyboardInterrupt:
+        tail_process.terminate()
+
+
+def current_process_is_orphaned(parent_pid):
+    parent_pid = int(parent_pid)
+    if sys.platform == 'win32':
+        import psutil
+
+        try:
+            parent = psutil.Process(parent_pid)
+            return parent.status() != psutil.STATUS_RUNNING
+        except psutil.NoSuchProcess:
+            return True
+
+    else:
+        return os.getppid() != parent_pid
 
 
 def _fileno(stream):