+++ python_modules/dagster/dagster_tests/core_tests/test_multiple_outputs.py
@@ -119,7 +119,7 @@ def test_multiple_outputs_only_emit_one():
 
     with pytest.raises(
         DagsterInvariantViolationError,
-        match='Try to get result for solid not_present in multiple_outputs_only_emit_one_pipeline. No such solid.',
+        match='Tried to get result for solid not_present in multiple_outputs_only_emit_one_pipeline. No such top level solid.',
     ):
         result.result_for_solid('not_present')