+++ CHANGES.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.6.2 (Upcoming)
+
+- Changed composition functions `@pipeline` and `@composite_solid` to automatically give solids
+  aliases with an incrementing integer suffix when there are conflicts. This removes to the need
+  to manually alias solid definitions that are used multiple times.
+- `execute_solid` test util now works on composite solids.
+- Added `--remote` flag to `dagster-graphql` for querying remote dagit servers.
+
 ## 0.6.1
 
 - Fixed an issue where Dagster public images tagged `latest` on Docker Hub were erroneously